+++ mux/src/stringutil.h
@@ -1,6 +1,6 @@
 // stringutil.h -- string utilities.
 //
-// $Id: stringutil.h,v 1.38 2004-06-10 15:23:32 sdennis Exp $
+// $Id: stringutil.h,v 1.39 2004-07-08 22:00:52 sdennis Exp $
 //
 // MUX 2.4
 // Copyright (C) 1998 through 2004 Solid Vertical Domains, Ltd. All
@@ -60,7 +60,7 @@ char *mux_strtok_parseLEN(MUX_STRTOK_STATE *tts, int *pnLen);
 char *mux_strtok_parse(MUX_STRTOK_STATE *tts);
 char *RemoveSetOfCharacters(char *pString, char *pSetToRemove);
 
-int mux_ltoa(long val, char *buf);
+size_t mux_ltoa(long val, char *buf);
 char *mux_ltoa_t(long val);
 void safe_ltoa(long val, char *buff, char **bufc);
 int mux_i64toa(INT64 val, char *buf);
